refactor(bigquery): use crate::model for storage_v1 types
