+++ tests/bigquery/src/writes/arrow.rs
@@ -19,7 +19,7 @@ use ::arrow::ipc::writer::StreamWriter;
 use ::arrow::record_batch::RecordBatch;
 use anyhow::Result;
 use google_cloud_bigquery::client::Write;
-use google_cloud_bigquery::write::model::{ArrowRecordBatch, ArrowSchema};
+use google_cloud_bigquery::model::{ArrowRecordBatch, ArrowSchema};
 use std::sync::Arc;
 
 pub async fn basic(
